IFATQCZZ-64 - Added Factory pattern to initialize WebDriver IFATQCZZ-13 - Added TestCase "As a Customer I want to see "Ordering" and "User Info" tab after login" With correct
