+++ src/test/java/academy/softserve/edu/enums/Drivers.java
@@ -4,6 +4,8 @@ package academy.softserve.edu.enums;
 public enum Drivers {
 
     CHROME,
-    SAFARI,
-    FIREFOX;
+    CHROME_MAC,
+    FIREFOX,
+    EXPLORER;
 }
+
