+++ src/test/java/academy/softserve/edu/pageobjects/UserInfoPage.java
@@ -11,8 +11,7 @@ public class UserInfoPage extends PageObject {
     public static final By MERCHANDISER_ORDERING_LINK = By.xpath("//ul[@id='nav']/descendant::a[@href='order.htm']");
     public static final By VISITOR_EXTRA_LINK = By.xpath("//ul[@id='nav']//a[1]");
     public static final String USER_INFO_PAGE_URL = "http://192.168.56.101:8080/oms5/userInfo.htm";
-    public static final By ORDERING_BUTTON = By.xpath("(.//ul[@id='nav']//a)[1]");
-    public static final By USER_INFO_BUTTON = By.xpath("(.//ul[@id='nav']//a)[2]");
+    public static final By USER_INFO_LINK = By.xpath("(.//ul[@id='nav']//a)[2]");
 
     public UserInfoPage(final WebDriver driver) {
         super(driver);
